MWPW-189465-component fix
