+++ test/blocks/submit-button/mocks/body.html
@@ -1,0 +1,5 @@
+<div class="submit-button">
+  <div>
+    <div><a href="https://partners.adobe.com/services/gravity/dxp/grant-download-access">Request access to Software Distribution</a></div>
+  </div>
+</div>
