--- test/blocks/text-with-submit/mocks/body.html
@@ -1,12 +1,0 @@
-<div class="text-with-submit">
-  <div>
-    <div>
-      <h2>Grant Access</h2>
-      <p>Use the button below to grant access to the portal.</p>
-      <p class="action-area">
-        <a href="/api/grant?submit=true">Grant Access</a>
-        <a href="/api/info">Learn More</a>
-      </p>
-    </div>
-  </div>
-</div>
